Create Migration.kt for old migration handling
